When a value is coerced, store the coerced value. fixes #16
Virtus coerces values for us. So if it see's a `"1"` for a `boolean`
attribute, it converts it to `true`.

The only issue with this is that Storext stores the original value
instead of the coerced value. So what is stored in the database & what
is returned by the model are different.

This change makes Storext save the same value as is returned by the
model.

It makes booleans work exactly like they do in Rails.

When a `"1"` or `"0"` is set from a form, it will be converted and
stored as `true` or `false`.